# Speed up `utils.diff.diff_values()`