--- a/astropy/utils/diff.py
+++ b/astropy/utils/diff.py
@@ -1,5 +1,6 @@
 import difflib
 import functools
+import math
 import sys
 from textwrap import indent
 
@@ -33,11 +34,11 @@ def diff_values(a, b, rtol=0.0, atol=0.0):
 
     """
     if isinstance(a, float) and isinstance(b, float):
-        if np.isnan(a) and np.isnan(b):
+        if math.isfinite(b):
+            return not abs(a - b) <= atol + rtol * abs(b)
+        if math.isnan(a) and math.isnan(b):
             return False
-        return not np.allclose(a, b, rtol=rtol, atol=atol)
-    else:
-        return a != b
+    return a != b
 
 
 def _ignore_astropy_terminal_size(func):