@eerovaher 2025-07-18
> Optimization should not change the result

What change are you talking about?